fix(imports): update lib/* import paths after folder refactor
