+++ src/app/(frontend)/layout.tsx
@@ -3,7 +3,7 @@ import type React from "react"
 import { Footer, Navbar } from "@/components/Composite"
 import "../globals.css"
 import type { Metadata, Viewport } from "next"
-import { getSession } from "@/lib/auth-session"
+import { getSession } from "@/lib/auth/auth-session"
 import { getDiscordWidgetData, getSocialLinks } from "@/lib/helpers"
 import { ApiRoutes, Routes } from "@/lib/routes"
 import { Providers } from "./providers"
