+++ src/app/(frontend)/profile/_components/MemberDetailsForm.tsx
@@ -3,7 +3,7 @@
 import { useState } from "react"
 import { ToggleableInput } from "@/components/Generic"
 import { Input, MultiSelect, Radio, Select } from "@/components/Primitive"
-import { ApiError } from "@/lib/api-client"
+import { ApiError } from "@/lib/api/api-client"
 import { toast } from "@/lib/toast"
 import type { Member } from "@/payload/payload-types"
 import { useUpdateMember } from "@/queries/useUpdateMember"
