+++ src/app/(frontend)/profile/_components/ProfilePageClient.tsx
@@ -3,7 +3,7 @@
 import { ArrowLeftEndOnRectangleIcon } from "@heroicons/react/24/solid"
 import type { User } from "better-auth"
 import { Button, Heading } from "@/components/Primitive"
-import { authClient } from "@/lib/auth-client"
+import { authClient } from "@/lib/auth/auth-client"
 import { useMember } from "@/queries/useMember"
 import { MemberDetailsForm } from "./MemberDetailsForm"
