+++ src/app/api/member/me/route.ts
@@ -1,6 +1,6 @@
 import { headers } from "next/headers"
 import { NextResponse } from "next/server"
-import { auth } from "@/lib/auth"
+import { auth } from "@/lib/auth/auth"
 import { AuthService } from "@/services/auth.service"
 
 const authService = new AuthService()
