+++ src/components/Composite/LoginForm/LoginForm.tsx
@@ -7,7 +7,7 @@ import { useState } from "react"
 import { useForm } from "react-hook-form"
 import { Button } from "@/components/Primitive"
 import { Input } from "@/components/Primitive/Input/Input"
-import { authClient } from "@/lib/auth-client"
+import { authClient } from "@/lib/auth/auth-client"
 import { Routes } from "@/lib/routes"
 import { toast } from "@/lib/toast"
 import { type LoginInput, type LoginOutput, loginSchema } from "@/types/schemas/login"
