+++ src/components/Composite/SignUpForm/EmailVerificationStep.tsx
@@ -6,8 +6,8 @@ import { useEffect, useRef, useState } from "react"
 import { Controller, useForm } from "react-hook-form"
 import { Button } from "@/components/Primitive"
 import { PinInput } from "@/components/Primitive/PinInput/PinInput"
-import { ApiError, api } from "@/lib/api-client"
-import { authClient } from "@/lib/auth-client"
+import { ApiError, api } from "@/lib/api/api-client"
+import { authClient } from "@/lib/auth/auth-client"
 import { ApiRoutes, Routes } from "@/lib/routes"
 import { toast } from "@/lib/toast"
 import {
