+++ src/components/Composite/SignUpForm/MemberStep.tsx
@@ -11,8 +11,8 @@ import { Button, Radio } from "@/components/Primitive"
 import { Input } from "@/components/Primitive/Input/Input"
 import { MultiSelect } from "@/components/Primitive/MultiSelect/MultiSelect"
 import { Select } from "@/components/Primitive/Select/Select"
-import { ApiError, api } from "@/lib/api-client"
-import { authClient } from "@/lib/auth-client"
+import { ApiError, api } from "@/lib/api/api-client"
+import { authClient } from "@/lib/auth/auth-client"
 import { ApiRoutes, Routes } from "@/lib/routes"
 import { toast } from "@/lib/toast"
 import { memberSchema } from "@/types/schemas/member"
