+++ src/components/Composite/SignUpForm/UserStep.tsx
@@ -6,7 +6,7 @@ import { useForm } from "react-hook-form"
 import type { z } from "zod"
 import { Button } from "@/components/Primitive"
 import { Input } from "@/components/Primitive/Input/Input"
-import { ApiError, api } from "@/lib/api-client"
+import { ApiError, api } from "@/lib/api/api-client"
 import { ApiRoutes } from "@/lib/routes"
 import { toast } from "@/lib/toast"
 import { createUserSchema } from "@/types/schemas/user"
