+++ src/payload.config.ts
@@ -7,7 +7,7 @@ import { lexicalEditor } from "@payloadcms/richtext-lexical"
 import { s3Storage } from "@payloadcms/storage-s3"
 import { buildConfig } from "payload"
 import sharp from "sharp"
-import { Slugs } from "./lib/slugs"
+import { Slugs } from "./lib/payload/slugs"
 import { Admin } from "./payload/collections/Admin"
 import { EmailVerificationCode } from "./payload/collections/EmailVerificationCode"
 import { Event } from "./payload/collections/Event"
