+++ src/queries/useMember.ts
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query"
-import { api } from "@/lib/api-client"
+import { api } from "@/lib/api/api-client"
 import { ApiRoutes } from "@/lib/routes"
 import type { Member } from "@/payload/payload-types"
 import { QueryKeys } from "./QueryKeys"
