+++ src/scripts/update-google-wallet-class.ts
@@ -1,6 +1,6 @@
-import { createPassClass } from "@/lib/wallet-basics"
-import { getPassClass } from "@/lib/wallet-designs"
-import { googleAuth, identity } from "@/lib/wallet-integration"
+import { createPassClass } from "@/lib/wallet/wallet-basics"
+import { getPassClass } from "@/lib/wallet/wallet-designs"
+import { googleAuth, identity } from "@/lib/wallet/wallet-integration"
 
 console.log(await createPassClass(googleAuth, getPassClass(identity)))
 console.log("... class updated")
